move some /train routes to new /filesystem router
