+++ api/src/controllers/train.controller.js
@@ -1,11 +1,7 @@
-const { promisify } = require('util');
 const fs = require('fs');
 const perf = require('execution-time')();
 const axios = require('axios');
 const { v4: uuidv4 } = require('uuid');
-const sharp = require('sharp');
-const { ExifTool } = require('exiftool-vendored');
-const sizeOf = promisify(require('image-size'));
 const { writer } = require('../util/fs.util');
 const database = require('../util/db.util');
 const sleep = require('../util/sleep.util');
@@ -16,128 +12,6 @@ const filesystem = require('../util/fs.util');
 const { respond, HTTPSuccess } = require('../util/respond.util');
 const { OK } = require('../constants/http-status');
 const { FRIGATE_URL, STORAGE_PATH, DETECTORS } = require('../constants');
-const { tryParseJSON } = require('../util/validators.util');
-
-module.exports.manage = async (req, res) => {
-  const { limit } = req.query;
-  let files = await filesystem.files().matches();
-  const width = req.query.width ? parseInt(req.query.width, 10) : 500;
-
-  const exiftool = new ExifTool({ taskTimeoutMillis: 1000 });
-
-  files = await Promise.all(
-    files.map(async (file) => {
-      const { UserComment: exif } = await exiftool.read(`${STORAGE_PATH}/${file.key}`);
-      const data = tryParseJSON(exif);
-      const { box, confidence, duration, detector } =
-        data !== false
-          ? data
-          : { confidence: null, duration: null, box: { width: null, height: null } };
-      const { birthtime: createdAt } = fs.statSync(`${STORAGE_PATH}/${file.key}`);
-      const base64 = await sharp(`${STORAGE_PATH}/${file.key}`).resize(width).toBuffer();
-
-      const { width: ogWidth, height: ogHeight } = await sizeOf(`${STORAGE_PATH}/${file.key}`);
-
-      const bbox = {
-        top: box.width ? `${(box.top / ogHeight) * 100}%` : 0,
-        width: box.width ? `${(box.width / ogWidth) * 100}%` : 0,
-        height: box.width ? `${(box.height / ogHeight) * 100}%` : 0,
-        left: box.width ? `${(box.left / ogWidth) * 100}%` : 0,
-      };
-
-      return {
-        ...file,
-        dimensions: box.width ? `${box.width}x${box.height}` : null,
-        bbox: box.width
-          ? `width: ${bbox.width}; height: ${bbox.height}; top: ${bbox.top}; left: ${bbox.left};`
-          : null,
-        confidence: confidence ? `${confidence}%` : null,
-        duration: duration ? `${duration} sec` : null,
-        detector: detector || null,
-        ago: time.ago(createdAt),
-        match: data.match,
-        type: data.type,
-        base64: base64.toString('base64'),
-      };
-    })
-  );
-  exiftool.end();
-
-  files = files.splice(0, limit);
-
-  // const rows = new Array(Math.ceil(files.length / 2)).fill().map(() => files.splice(0, 2));
-  const folders = await filesystem.folders().train();
-  return respond(HTTPSuccess(200, { files, folders }), res);
-};
-
-module.exports.file = () => {
-  return {
-    delete: (req, res) => {
-      try {
-        const files = req.body;
-        files.forEach((file) => {
-          filesystem.delete(`${STORAGE_PATH}/${file.key}`);
-        });
-        respond(HTTPSuccess(OK, { sucess: true }), res);
-      } catch (error) {
-        logger.log(`manage delete error: ${error.message}`);
-        respond(error, res);
-      }
-    },
-    move: async (req, res) => {
-      try {
-        const { folder, files } = req.body;
-
-        files.forEach((file) => {
-          filesystem.move(
-            `${STORAGE_PATH}/${file.key}`,
-            `${STORAGE_PATH}/train/${folder}/${file.filename}`
-          );
-        });
-        database.insert('init', await filesystem.files().train());
-        respond(HTTPSuccess(OK, { sucess: true }), res);
-        await train.queue(database.files('untrained', folder));
-      } catch (error) {
-        logger.log(`manage move error: ${error.message}`);
-        respond(error, res);
-      }
-    },
-  };
-};
-
-module.exports.delete = async (req, res) => {
-  try {
-    const { name } = req.params;
-
-    perf.start();
-    const promises = [];
-
-    DETECTORS.forEach((detector) => {
-      promises.push(train.remove({ detector, name }));
-      const db = database.connect();
-      db.prepare('DELETE FROM train WHERE detector = ? AND name = ?').run(detector, name);
-    });
-    const results = await Promise.all(promises);
-
-    logger.log(
-      `${time.current()}\ndone untraining for ${name} in ${parseFloat(
-        (perf.stop().time / 1000).toFixed(2)
-      )} sec`
-    );
-
-    const output = results.map((result, i) => {
-      return {
-        detector: DETECTORS[i],
-        results: result,
-      };
-    });
-
-    respond(HTTPSuccess(OK, output), res);
-  } catch (error) {
-    logger.log(`train delete error: ${error.message}`);
-    respond(error, res);
-  }
-};
 
 module.exports.camera = async (req, res) => {
   const { name, camera } = req.params;
@@ -176,6 +50,40 @@ module.exports.camera = async (req, res) => {
     respond(HTTPSuccess(OK, { camera, name, results }), res);
   } catch (error) {
     logger.log(`train camera error: ${error.message}`);
+    respond(error, res);
+  }
+};
+
+module.exports.delete = async (req, res) => {
+  try {
+    const { name } = req.params;
+
+    perf.start();
+    const promises = [];
+
+    DETECTORS.forEach((detector) => {
+      promises.push(train.remove({ detector, name }));
+      const db = database.connect();
+      db.prepare('DELETE FROM train WHERE detector = ? AND name = ?').run(detector, name);
+    });
+    const results = await Promise.all(promises);
+
+    logger.log(
+      `${time.current()}\ndone untraining for ${name} in ${parseFloat(
+        (perf.stop().time / 1000).toFixed(2)
+      )} sec`
+    );
+
+    const output = results.map((result, i) => {
+      return {
+        detector: DETECTORS[i],
+        results: result,
+      };
+    });
+
+    respond(HTTPSuccess(OK, output), res);
+  } catch (error) {
+    logger.log(`train delete error: ${error.message}`);
     respond(error, res);
   }
 };
